Abort optimize script if imageOptim fails Move example redux modules to /examples nested reducer
