+++ src/redux/modules/index.js
@@ -2,17 +2,6 @@
 import { combineReducers } from 'redux'
 import { reducer as form } from 'redux-form'
 
-import env from 'src/redux/modules/env'
-import request from 'src/redux/modules/request'
-import init from 'src/redux/modules/init'
-import auth from 'src/redux/modules/auth'
-import ui from 'src/redux/modules/ui'
-
-// Examples
-import profile from 'src/redux/modules/profile'
-import todos from 'src/redux/modules/todos'
-
-
 // Types
 import type { EnvState } from 'config/env'
 import type { RequestState } from 'src/redux/modules/request'
@@ -21,8 +10,16 @@ import type { AuthState } from 'src/redux/modules/auth'
 import type { UIState } from 'src/redux/modules/ui'
 
 // Examples
-import type { ProfileState } from 'src/redux/modules/profile'
-import type { TodosState } from 'src/redux/modules/todos'
+import type { ExamplesState } from 'src/redux/modules/examples'
+
+import env from './env'
+import request from './request'
+import init from './init'
+import auth from './auth'
+import ui from './ui'
+
+// Examples
+import examples from './examples'
 
 
 export type RootReducerState = {
@@ -33,8 +30,7 @@ export type RootReducerState = {
   ui: UIState,
 
   // Examples
-  profile: ProfileState,
-  todos: TodosState,
+  examples: ExamplesState,
 
   // redux-form
   form: Object,
@@ -50,8 +46,7 @@ const rootReducer = combineReducers({
   ui,
 
   // Examples
-  profile,
-  todos,
+  examples,
 
   // redux-form
   form,
