+++ src/selectors/areWeLoading.js
@@ -4,31 +4,35 @@ import type { RootReducerState } from 'src/redux/modules'
 
 const getAuthFetching = (state: RootReducerState) => state.auth.isFetching
 const getInitLoading = (state: RootReducerState) => state.init.isLoading
-const getProfileFetching = (state: RootReducerState) => state.profile.isFetching
-const getTodosFetching = (state: RootReducerState) => state.todos.isFetching
 const getUiLoading = (state: RootReducerState) => state.ui.showLoading
+
+// Examples
+const getProfileFetching = (state: RootReducerState) =>
+  state.examples.profile.isFetching
+const getTodosFetching = (state: RootReducerState) =>
+  state.examples.todos.isFetching
 
 
 const areWeLoading = createSelector([
   getAuthFetching,
   getInitLoading,
+  getUiLoading,
   getProfileFetching,
   getTodosFetching,
-  getUiLoading,
 ], (
   authFetching,
   initLoading,
+  uiLoading,
   profileFetching,
   todosFetching,
-  uiLoading,
 ) => {
 
   if (
     authFetching
     || initLoading
+    || uiLoading
     || profileFetching
     || todosFetching
-    || uiLoading
   ) {
 
     return true
